Replace JSON array with map for cmd PUT params
The current implementation only accepts a JSON array
of maps for the command parms pass as the request body.
As this is wrong (see issue #141), this change modifies
the code to accept a single map of parameters.
To keep device-simple working, the PUT command schedule
event params is also modified.
fix #141

Signed-off-by: Cloud Tsai <cloudxxx8@gmail.com>